requirements.txt
aiohttp==3.10.10
azure-identity==1.19.0
azure-servicebus==7.12.3